Added the popup menu for sorting.
No functionality, but the menu is solid.

INCLUDING generating the "By title" icon (since the desktop Nitro used text instead of a PNG).

I probably need to remove a ton of the animations that are not used,

The library I modified is released under the Apache license, so I probably need to get that in here too.