Switched to tablet compatibility.
Instead of having two Activities (one for Lists and one for Tasks), there is now just a List activity that creates a Task
object.

The main reason for this is tablet compatibility. On tablets, I would like to have it almost exactly like the desktop
version, i.e., lists on the left and tasks on the right. Having it as two activities would mean that I'd practically have to
re-write it, but instead, if I just change it from a ViewFlipper for phones to a LinearLayout for tablets, the code should
still work perfectly, having only to comment out the parts where it is told to flip views.

That was a long explanation for a git comment. I'm going to keep talking. Blah blah blah, blah blah.