Started sync work. NOTE: BROKEN BUILD.
(I realize it's not proper to make a commit where the app is broken, but I need to push my changes to save them for the night.)

This is mostly the sync.coffee file ported to Java. But there are parts I don't quite understand yet, so I left those parts in Coffeescript as markers.

I also left some comments at the bottom of SyncHelper_v2 about what I believe still needs to be done. A talk with George will hopefully clear it up.